backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.7
sqlalchemy==2.0.27
python-dotenv==1.0.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0
xlrd==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2